frontal_prod/antropometrico/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.24.3
torch==2.0.1+cu118
torchvision==0.15.2+cu118
dlib==19.22.1
matplotlib==3.7.2
Pillow==10.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
--extra-index-url https://download.pytorch.org/whl/cu118
